% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{create_docker}
\alias{create_docker}
\title{Create an Instance with Docker}
\usage{
create_docker(client, projectId, name, regionVM, typeVM, sshKeyId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The Id of the project}

\item{name}{The name of the instance}

\item{regionVM}{The name of datacenter: SBG3, BHS3, WAW1, UK1, DE1, GRA3}

\item{typeVM}{The range of your cloud server ( s1-2, s1-4: https://www.ovh.co.uk/public-cloud/instances/prices/)}

\item{sshKeyId}{The Id of ssh key}
}
\value{
JSON
}
\description{
Create an Instance with Docker
}
\details{
Logic Workflow
1. get the available flavor where the name matching with the specific region
2. get the active image where the name matching with "Docker"
3. create the instance
}

