% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{create_instance}
\alias{create_instance}
\title{Create new Instance}
\usage{
create_instance(client, projectId, flavorId, imageId, name, region, sshKeyId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The id of your cloud project}

\item{flavorId}{The id of your flavor}

\item{imageId}{The id of your distribution}

\item{name}{The name of your instance}

\item{region}{The name of your datacenter}

\item{sshKeyId}{The Id of your SSH Key}
}
\value{
JSON
}
\description{
\url{https://api.ovh.com/console/#/cloud}
}

