% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{create_key}
\alias{create_key}
\title{Create SSH Key}
\usage{
create_key(client, projectId, name, pathPublicKey)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The id of the cloud project}

\item{name}{The name of the SSH Key}

\item{pathPublicKey}{The path of your public SSH Key}
}
\value{
JSON
}
\description{
\url{https://api.ovh.com/console/#/cloud}
}

