% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{create_project}
\alias{create_project}
\title{Create Cloud Project}
\usage{
create_project(client, name)
}
\arguments{
\item{client}{The client previously loaded}

\item{name}{The name of the project}
}
\value{
JSON
}
\description{
Start a new cloud project
}
\details{
\url{https://api.ovh.com/console/#/cloud}
}

