% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{delete_instance}
\alias{delete_instance}
\title{Delete an instance}
\usage{
delete_instance(client, projectId, instanceId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The Id of the project}

\item{instanceId}{The Id of the instance}
}
\description{
\url{https://api.ovh.com/console/#/cloud}
}

