% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_docker_file}
\alias{get_docker_file}
\title{Show the docker template files}
\usage{
get_docker_file(path)
}
\arguments{
\item{path}{The file name you want

This returns the file location of template files for use Docker}
}
\value{
file location
}
\description{
Show the docker template files
}

