% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_flavors}
\alias{get_flavors}
\title{Get Flavors}
\usage{
get_flavors(client, projectId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The id of the project}
}
\value{
Data.frame
}
\description{
Get all flavors
}
\details{
\url{https://api.ovh.com/console/#/cloud}


In OpenStack, a flavor defines the compute, memory, and storage capacity of a virtual server, also known as an instance.
}

