% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_images}
\alias{get_images}
\title{Get Images}
\usage{
get_images(client, projectId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The id of the instance}
}
\value{
Data.frame
}
\description{
Get all images
}
\details{
\url{https://api.ovh.com/console/#/cloud}


A virtual machine image is a single file that contains a virtual disk that has a bootable operating system installed on it.
Images are used to create virtual machine instances within the cloud.
}

