% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_intance}
\alias{get_intance}
\title{Get instance Detail from a specific project}
\usage{
get_intance(client, projectId, instanceId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The id of the project}

\item{instanceId}{The id of the instance}
}
\value{
JSON
}
\description{
\url{https://api.ovh.com/console/#/cloud}
}

