% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_intances}
\alias{get_intances}
\title{Get instances from a specific project}
\usage{
get_intances(client, projectId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The Id of the project}
}
\value{
JSON
}
\description{
\url{https://api.ovh.com/console/#/cloud}
}

