% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_keys}
\alias{get_keys}
\title{Get SSH Keys}
\usage{
get_keys(client, projectId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The Id of the project}
}
\value{
JSON
}
\description{
\url{https://api.ovh.com/console/#/cloud}
}

