% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_project_info}
\alias{get_project_info}
\title{Get Project Info}
\usage{
get_project_info(client, projectId)
}
\arguments{
\item{client}{The client previously loaded}

\item{projectId}{The Id of the project}
}
\value{
JSON
}
\description{
Get this object properties
}
\details{
\url{https://api.ovh.com/console/#/cloud}
}

