% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{get_projects}
\alias{get_projects}
\title{Get Projects}
\usage{
get_projects(client)
}
\arguments{
\item{client}{The client previously loaded}
}
\value{
JSON
}
\description{
List available services
}
\details{
\url{https://api.ovh.com/console/#/cloud}
}

