% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkler.R
\name{get_sparkler_file}
\alias{get_sparkler_file}
\title{Show the sparkler-config template files}
\usage{
get_sparkler_file(template)
}
\arguments{
\item{template}{This returns the file location of template files for use Docker}
}
\value{
file location
}
\description{
Show the sparkler-config template files
}

