% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{import_ovh}
\alias{import_ovh}
\title{Lightweight wrapper around OVH's APIs. Handles all the hard work including credential creation and requests signing.
Import Python Library : https://github.com/ovh/python-ovh}
\usage{
import_ovh()
}
\description{
Lightweight wrapper around OVH's APIs. Handles all the hard work including credential creation and requests signing.
Import Python Library : https://github.com/ovh/python-ovh
}
\details{
An Instance resource.
1. you need to install Python
2. execute this command : pip install ovh
}

