% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{Instance}
\alias{Instance}
\title{Instance Object}
\usage{
Instance(name = NULL, machineType = NULL, region = NULL,
  description = NULL, projectId = NULL, instanceId = NULL, ip = NULL)
}
\arguments{
\item{name}{The name of the resource, provided by the client when initially creating the resource}

\item{machineType}{The type of your server ( s1-2, s1-4 )}

\item{region}{The name of datacenter : SBG3, BHS3, WAW1, UK1, DE1, GRA3}

\item{description}{( not used )}

\item{projectId}{The id of the cloud project}

\item{instanceId}{The id of the instance}

\item{ip}{The IP of your instance ( format IPV4 )}
}
\value{
Instance object
}
\description{
Instance Object
}
\details{
An Instance resource.

About regions :
SBG3 Datacenter is in France
BHS3 Datacenter is in Canada
WAW1 Datacenter is in Poland
UK1  Datacenter is in UK
DE1  Datacenter is in Deutch
GRA3 Datacenter is in France

About range of cloud servers
https://www.ovh.co.uk/public-cloud/instances/prices/
}
\keyword{internal}

