% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh.R
\name{load_client}
\alias{load_client}
\title{Create a client using configuration}
\usage{
load_client(ovh, endpoint, appkey, appsecret, consumerkey)
}
\arguments{
\item{ovh}{The OVH lib}

\item{endpoint}{e.g ; ovh-eu}

\item{appkey}{Application key}

\item{appsecret}{Secret app key}

\item{consumerkey}{Consumer key}
}
\value{
client
}
\description{
Create a client using configuration
}
\details{
You need to get appkey, appsecret, consumerkey
The four last can be took directly from their page createApp by filling fields. It is easy.
https://api.ovh.com/createToken/index.cgi?GET=/*&POST=/*&PUT=/*&DELETE=/*
}

