% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkler.R
\name{sparkler.check}
\alias{sparkler.check}
\title{Check State}
\usage{
sparkler.check(vm, pattern, crawlid, maxUrls)
}
\arguments{
\item{vm}{The Instance object}

\item{pattern}{The hostname.}

\item{crawlid}{The Id of crawl.}

\item{maxUrls}{The max URLs for fetching.}
}
\value{
Boolean
}
\description{
Print the state of crawl
}
\details{
Display progress bar
Return TRUE When UNFECTCHED count becomes zero
}
\examples{

\dontrun{

library(RsparkleR)
ovh <- import_ovh()
client <- load_client(ovh, endpoint, application_key, application_secret, consumer_key)

sshPubKeyPath  <- 'C:/Users/vterrasi/.ssh/id_rsa.pub'
sshPrivKeyPath <- 'C:/Users/vterrasi/.ssh/id_rsa'

vm <- sparkler.create(client, regionVM="UK1", typeVM="s1-4", sshPubKeyPath, sshPrivKeyPath)

url <- "https://www.YOUR WEBSITE.com"
pattern <- "www.YOUR WEBSITE.com"
topN <- 1000

crawlid <- sparkler.crawl(vm, url, topN, topGroups, maxIter, debug, mode="fast")

checkAll <- sparkler.check(vm, pattern, crawlid, topN)

}

}

