% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkler.R
\name{sparkler.create}
\alias{sparkler.create}
\title{Create an Instance with Sparkler}
\usage{
sparkler.create(client, regionVM, typeVM, sshPubKeyPath, sshPrivKeyPath)
}
\arguments{
\item{client}{The name of the resource, provided by the client when initially creating the resource}

\item{regionVM}{The name of datacenter: SBG3, BHS3, WAW1, UK1, DE1, GRA3}

\item{typeVM}{The range of your cloud server ( s1-2, s1-4: https://www.ovh.co.uk/public-cloud/instances/prices/)}

\item{sshPubKeyPath}{The path of public SSH key}

\item{sshPrivKeyPath}{The path of private SSH key}
}
\value{
A \code{instance} object
}
\description{
Create an Instance with Sparkler
}
\details{
Make sure the instance is big enough to handle what you need,
  for instance the default \code{s1-4} will hang the instance when trying to install large R libraries.
}
\section{Creation logic}{


You need these parameters defined to call the right function for creation.  Check the function definitions for more details.

If the Instance name exists but is not running, it starts the Instance and return the Instance object
If the Instance doesn't exist, it creates the Instance object by using the createDocker function.
If the Instance is running, it will return the Instance object
}
\examples{

\dontrun{

library(RsparkleR)

ovh <- import_ovh()
client <- load_client(ovh, endpoint, application_key, application_secret, consumer_key)

sshPubKeyPath  <- 'C:/Users/vterrasi/.ssh/id_rsa.pub'
sshPrivKeyPath <- 'C:/Users/vterrasi/.ssh/id_rsa'

vm <- sparkler.create(client, regionVM="UK1", typeVM="s1-4", sshPubKeyPath, sshPrivKeyPath)

}

}

