% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkler.R
\name{sparkler.read.csv}
\alias{sparkler.read.csv}
\title{Read Solr CSV}
\usage{
sparkler.read.csv(vm, pattern, crawl_id, limit, extracted = FALSE)
}
\arguments{
\item{vm}{The Instance object}

\item{pattern}{The hostname}

\item{crawl_id}{The Id of crawl}

\item{limit}{The max URLs you want}

\item{extracted}{It TRUE, will get extracted text}
}
\value{
A dataframe
}
\description{
Read Solr CSV
}
\details{
You can find a SolR tutorial here : https://lucene.apache.org/solr/guide/7_2/solr-tutorial.html
}
\examples{

\dontrun{

library(RsparkleR)

ovh <- import_ovh()
client <- load_client(ovh,endpoint,application_key,application_secret,consumer_key)

sshPubKeyPath  <- 'C:/Users/vterrasi/.ssh/id_rsa.pub'
sshPrivKeyPath <- 'C:/Users/vterrasi/.ssh/id_rsa'

vm <- sparkler.create(client, regionVM="UK1", typeVM="s1-4", sshPubKeyPath, sshPrivKeyPath)

topN <- 1000

crawlid <- sparkler.crawl(vm, url, topN, topGroups, maxIter, debug=FALSE, mode="fast")

crawlDF <- sparkler.read.csv(vm, pattern, crawlid, topN, extracted=FALSE)

}

}

