% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkler.R
\name{sparkler.start}
\alias{sparkler.start}
\title{Start a docker with Sparkler}
\usage{
sparkler.start(vm, debug = FALSE)
}
\arguments{
\item{vm}{The Instance object}

\item{debug}{If TRUE, will see debug messages.}
}
\description{
Start a docker with Sparkler
}
\details{
Check if Docker exists and running
   If not, we create the docker with Sparkler
   If exists, we restart it
}
\examples{

\dontrun{

library(RsparkleR)
ovh <- import_ovh()
client <- load_client(ovh,endpoint,application_key,application_secret,consumer_key)

sshPubKeyPath  <- 'C:/Users/vterrasi/.ssh/id_rsa.pub'
sshPrivKeyPath <- 'C:/Users/vterrasi/.ssh/id_rsa'

vm <- sparkler.create(client, regionVM="UK1", typeVM="s1-4", sshPubKeyPath, sshPrivKeyPath)

sparkler.start(vm, debug=FALSE)

}

}

