% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkler.R
\name{sparkler.stop}
\alias{sparkler.stop}
\title{Shutdown Solr and remove the container}
\usage{
sparkler.stop(vm, debug = FALSE)
}
\arguments{
\item{vm}{The Instance object}

\item{debug}{If TRUE, will see debug messages.}
}
\description{
Shutdown Solr and remove the container
}
\examples{

\dontrun{

library(RsparkleR)

ovh <- import_ovh()
client <- load_client(ovh, endpoint, application_key, application_secret, consumer_key)

sshPubKeyPath  <- 'C:/Users/vterrasi/.ssh/id_rsa.pub'
sshPrivKeyPath <- 'C:/Users/vterrasi/.ssh/id_rsa'

vm <- sparkler.create(client, regionVM="UK1", typeVM="s1-4", sshPubKeyPath, sshPrivKeyPath)

sparkler.stop(vm, debug=FALSE)

}

}

